+++ tests/runners/test_model_upload.py
@@ -10,13 +10,20 @@ import yaml
 from clarifai_grpc.grpc.api.status import status_code_pb2
 
 from clarifai.client import User
-from clarifai.runners.models.model_builder import ModelBuilder
+from clarifai.runners.models.model_builder import ModelBuilder, deploy_model
 
 MODEL_PATH = os.path.join(os.path.dirname(__file__), "dummy_runner_models")
 CLARIFAI_USER_ID = os.environ["CLARIFAI_USER_ID"]
 CLARIFAI_PAT = os.environ["CLARIFAI_PAT"]
 NOW = uuid.uuid4().hex[:10]
 CREATE_APP_ID = f"pytest-model-upload-test-{NOW}"
+CREATE_COMPUTE_CLUSTER_ID = f"test-compute_cluster-{NOW}"
+CREATE_NODEPOOL_ID = f"test-nodepool-{NOW}"
+CREATE_DEPLOYMENT_ID = f"test-deployment-{NOW}"
+COMPUTE_CLUSTER_CONFIG_FILE = (
+    "tests/compute_orchestration/configs/example_compute_cluster_config.yaml"
+)
+NODEPOOL_CONFIG_FILE = "tests/compute_orchestration/configs/example_nodepool_config.yaml"
 
 
 def check_app_exists():
@@ -51,6 +58,15 @@ def create_app():
         print(f"Creating app '{CREATE_APP_ID}'...")
         user.create_app(app_id=CREATE_APP_ID)
     return CREATE_APP_ID, user
+
+
+@pytest.fixture
+def client():
+    return User(
+        user_id=CLARIFAI_USER_ID,
+        base_url=os.environ.get('CLARIFAI_API_BASE', 'https://api.clarifai.com'),
+        pat=CLARIFAI_PAT,
+    )
 
 
 @pytest.fixture(scope="module")
@@ -120,7 +136,7 @@ def test_init_valid_folder(model_builder):
     assert "config.yaml" in os.listdir(model_builder.folder)
 
 
-def test_model_uploader_flow(dummy_models_path):
+def test_model_uploader_flow(dummy_models_path, client):
     """
     End-to-end test that:
     1. Initializes the ModelBuilder on the dummy_runner_models folder
@@ -128,6 +144,8 @@ def test_model_uploader_flow(dummy_models_path):
     3. Creates or reuses an existing model
     4. Uploads a new model version
     5. Waits for the build
+    6. Deploy the model to test
+    7. Delete the deployment
     """
     # Initialize
     builder = ModelBuilder(folder=str(dummy_models_path))
@@ -167,6 +185,38 @@ def test_model_uploader_flow(dummy_models_path):
     assert builder.model_version_id is not None, "Model version upload failed to initialize"
 
     print(f"Test completed successfully with model_version_id={builder.model_version_id}")
+
+    model = builder.config.get('model')
+
+    # Setup compute cluster and nodepool
+    compute_cluster = client.create_compute_cluster(
+        compute_cluster_id=CREATE_COMPUTE_CLUSTER_ID,
+        config_filepath=COMPUTE_CLUSTER_CONFIG_FILE,
+    )
+    nodepool = compute_cluster.create_nodepool(
+        nodepool_id=CREATE_NODEPOOL_ID, config_filepath=NODEPOOL_CONFIG_FILE
+    )
+
+    success = deploy_model(
+        model_id=model.get('id'),
+        app_id=model.get('app_id'),
+        user_id=model.get('user_id'),
+        deployment_id=CREATE_DEPLOYMENT_ID,
+        model_version_id=builder.model_version_id,
+        nodepool_id=CREATE_NODEPOOL_ID,
+        compute_cluster_id=CREATE_COMPUTE_CLUSTER_ID,
+        cluster_user_id=CLARIFAI_USER_ID,
+        min_replicas=1,
+    )
+
+    assert success is True, "Model deployment failed"
+
+    print(f"Test completed successfully with deployment_id={CREATE_DEPLOYMENT_ID}")
+
+    # Clean up resources
+    nodepool.delete_deployments([CREATE_DEPLOYMENT_ID])
+    compute_cluster.delete_nodepools([CREATE_NODEPOOL_ID])
+    client.delete_compute_clusters([CREATE_COMPUTE_CLUSTER_ID])
 
 
 @pytest.fixture